pyproject.toml
[project]
name = "fraud-detection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.7.0",
    "mlflow>=2.3.0",
    "numpy>=1.23.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.2.0",
    "seaborn>=0.12.0",
    "umap-learn>=0.5.5",
]
